refactor: network files and asset model
